Disable strace test execution
Due to some docker capability restrictions (e.g.: lack of
cap_sys_ptrace) around 20 tests are failing per build and per
architecture that prevent the strace to be successfully build.

Since the strace considered not to be a production package but used only
for troubleshooting purposes we can live without these tests.